@bruw/101/enh/organization of action folders (#102)
* organization of resources into more specific folders

* Test organizanization
